refactor(core/revision): isolate trash revision in dedicated controller
