+++ EMS/core-bundle/src/Resources/config/controllers.xml
@@ -382,6 +382,13 @@
             <tag name="container.service_subscriber"/>
             <tag name="controller.service_arguments"/>
         </service>
+        <service id="EMS\CoreBundle\Controller\Revision\TrashController" public="true">
+            <argument type="service" id="ems.service.data" />
+            <argument>%ems_core.template_namespace%</argument>
+            <call method="setContainer"/>
+            <tag name="container.service_subscriber"/>
+            <tag name="controller.service_arguments"/>
+        </service>
 
         <!-- Search -->
         <service id="EMS\CoreBundle\Controller\Search\QuerySearchController" public="true">
